+++ src/main/java/com/cnzk/pojo/TbRole.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public class TbRole {
 
-  private Integer roleId;
+  private Long roleId;
   private String roleName;
   private Integer roleSort;
   private List<TbMenu> menutblList;
@@ -13,23 +13,23 @@ public class TbRole {
   public TbRole() {
   }
 
-  public TbRole(Integer roleId, String roleName) {
+  public TbRole(Long roleId, String roleName) {
     this.roleId = roleId;
     this.roleName = roleName;
   }
 
-  public TbRole(Integer roleId, String roleName, Integer roleSort, List<TbMenu> menutblList) {
+  public TbRole(Long roleId, String roleName, Integer roleSort, List<TbMenu> menutblList) {
     this.roleId = roleId;
     this.roleName = roleName;
     this.roleSort = roleSort;
     this.menutblList = menutblList;
   }
 
-  public Integer getRoleId() {
+  public Long getRoleId() {
     return roleId;
   }
 
-  public void setRoleId(Integer roleId) {
+  public void setRoleId(Long roleId) {
     this.roleId = roleId;
   }
